Refactored message classes package location
